+++ src/app/exhibits/page.tsx
@@ -10,6 +10,7 @@ import { DocumentResponseDto } from "../../types/types";
 import { IconExternalLink } from "@tabler/icons-react";
 import { Viewer } from "@react-pdf-viewer/core";
 import "@react-pdf-viewer/core/lib/styles/index.css";
+import { Layout as BaseLayout } from "@components/layout";
 
 export default function BlogPostList() {
   const { push } = useNavigation();
@@ -46,55 +47,57 @@ export default function BlogPostList() {
   }, [documentData]);
 
   return (
-    <Worker workerUrl="https://unpkg.com/pdfjs-dist@3.4.120/build/pdf.worker.min.js">
-      <div className="p-5 h-full flex flex-col">
-        <div className="text-3xl text-black">
-          {caseData?.data.title}
-          <span className="text-sm ml-3 mr-1">/</span>
-          <span className="text-sm">All documents</span>
-        </div>
-        <div className="relative pt-4">
-          <LoadingOverlay
-            visible={caseLoading || documentLoading}
-            zIndex={1000}
-            overlayProps={{ radius: "sm", blur: 2 }}
-            loaderProps={{ color: "pink", type: "bars" }}
-          />
-        </div>
-        <div className="grid grid-cols-10 flex-1">
-          <div className="col-span-4 relative">
-            {documents
-              .filter((doc) => doc.type === DocType.EXHIBIT)
-              .map((doc, _i) => (
-                <div
-                  key={doc.id}
-                  className={`border p-4 border-r-0 flex justify-between items-center cursor-pointer
+    <BaseLayout>
+      <Worker workerUrl="https://unpkg.com/pdfjs-dist@3.4.120/build/pdf.worker.min.js">
+        <div className="p-5 h-full flex flex-col">
+          <div className="text-3xl text-black">
+            {caseData?.data.title}
+            <span className="text-sm ml-3 mr-1">/</span>
+            <span className="text-sm">All documents</span>
+          </div>
+          <div className="relative pt-4">
+            <LoadingOverlay
+              visible={caseLoading || documentLoading}
+              zIndex={1000}
+              overlayProps={{ radius: "sm", blur: 2 }}
+              loaderProps={{ color: "pink", type: "bars" }}
+            />
+          </div>
+          <div className="grid grid-cols-10 flex-1">
+            <div className="col-span-4 relative">
+              {documents
+                .filter((doc) => doc.type === DocType.EXHIBIT)
+                .map((doc, _i) => (
+                  <div
+                    key={doc.id}
+                    className={`border p-4 border-r-0 flex justify-between items-center cursor-pointer
                   ${
                     selectedMainDocument?.id === doc.id
                       ? "text-[#3040d6]"
                       : "text-[#6e6e6e]"
                   } 
                   ${_i === 0 ? "border-t" : "border-t-0"}`}
-                >
-                  <div
-                    onClick={() => setSelectedMainDocument(doc)}
-                    className="font-bold flex gap-2 items-center cursor-pointer text-sm"
                   >
-                    <IconExternalLink
-                      style={{ width: rem(20), height: rem(20) }}
-                    />
-                    {doc.title}
+                    <div
+                      onClick={() => setSelectedMainDocument(doc)}
+                      className="font-bold flex gap-2 items-center cursor-pointer text-sm"
+                    >
+                      <IconExternalLink
+                        style={{ width: rem(20), height: rem(20) }}
+                      />
+                      {doc.title}
+                    </div>
                   </div>
-                </div>
-              ))}
-          </div>
-          <div className="border col-span-6 p-4">
-            {selectedMainDocument && (
-              <Viewer fileUrl={selectedMainDocument.mediaUrl} />
-            )}
+                ))}
+            </div>
+            <div className="border col-span-6 p-4">
+              {selectedMainDocument && (
+                <Viewer fileUrl={selectedMainDocument.mediaUrl} />
+              )}
+            </div>
           </div>
         </div>
-      </div>
-    </Worker>
+      </Worker>
+    </BaseLayout>
   );
 }
