--- src/utils/menuData.ts
@@ -1,17 +1,0 @@
-// src/utils/menuData.ts
-export const menuItems = [
-  {
-    key: "documents",
-    label: "Documents",
-    route: "/documents",
-  },
-  {
-    key: "exhibits",
-    label: "Exhibit Library",
-    route: "/exhibits",
-  },
-  // Add more items as needed
-];
-
-// src/utils/menuData.ts
-export const dashboardMenuItems = [];
